Allow new primary EEPs to be added
New EEPs can be added via `custom_install.py` or directly to `to_eep()` by assigning an ordered list of EEP function key strings to `eep_order`.

`eep_order = ['prems', 'zams', ...]`

I also fixed an issue with calc_HZ, where the ZAMS EEP was set to the size of the first interval, not the size of the first interval +1.